requirements.txt
appdirs==1.4.3
appnope==0.1.0
asgi-redis==1.3.0
asgiref==1.1.1
attrs==16.3.0
autobahn==0.18.2
Automat==0.5.0
channels==1.1.3
constantly==15.1.0
daphne==1.2.0
decorator==4.0.11
Django==1.11
django-admin==1.0.2
django-appconf==1.0.2
django-debug-panel==0.8.3
django-debug-toolbar==1.7
django-filter==1.0.1
djangorestframework==3.6.2
flake8==3.3.0
incremental==16.10.1
ipdb==0.10.2
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
Markdown==2.6.8
mccabe==0.6.1
msgpack-python==0.4.8
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.0
pip==9.0.1
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
redis==2.10.5
setuptools==35.0.2
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
stringcase==1.0.6
traitlets==4.3.2
Twisted==17.1.0
txaio==2.7.0
wcwidth==0.1.7
wheel==0.29.0
zope.interface==4.4.0
